Extract array modules
