+++ cpp/hello/tests/test_hello_c.cpp
@@ -2,26 +2,7 @@
 
 #include <hello.h>
 
-TEST_CASE("[C] to_uppercase()")
-{
-  char *string = hello__string__to_uppercase("Hello World!");
-
-  CHECK(std::string(string) == "HELLO WORLD!");
-
-  hello__string__destroy_string(string);
-}
-
-TEST_CASE("[C] to_uppercase_safe()")
-{
-  std::string in_string("Hello World!");
-  std::string out_string(in_string.length(), '\0');
-
-  hello__string__to_uppercase_safe(in_string.c_str(), (char *)out_string.c_str());
-
-  REQUIRE(out_string == "HELLO WORLD!");
-}
-
-TEST_CASE("[C] concat()")
+TEST_CASE("[C] hello__array__concat()")
 {
   float a[] = {1.f, 2.f};
   size_t a_length = sizeof(a) / sizeof(a[0]);
@@ -29,7 +10,7 @@ TEST_CASE("[C] concat()")
   float b[] = {3.f, 4.f, 5.f};
   size_t b_length = sizeof(b) / sizeof(b[0]);
 
-  float *array = hello__concat(a, a_length, b, b_length);
+  float *array = hello__array__concat(a, a_length, b, b_length);
 
   CHECK(array[0] == 1.f);
   CHECK(array[1] == 2.f);
@@ -37,10 +18,10 @@ TEST_CASE("[C] concat()")
   CHECK(array[3] == 4.f);
   CHECK(array[4] == 5.f);
 
-  hello__destroy_array(array);
+  hello__array__destroy_array(array);
 }
 
-TEST_CASE("[C] concat_safe()")
+TEST_CASE("[C] hello__array__concat_safe()")
 {
   float a[] = {1.f, 2.f};
   const size_t a_length = sizeof(a) / sizeof(a[0]);
@@ -50,7 +31,7 @@ TEST_CASE("[C] concat_safe()")
 
   float c[a_length + b_length];
 
-  hello__concat_safe(a, a_length, b, b_length, c);
+  hello__array__concat_safe(a, a_length, b, b_length, c);
 
   REQUIRE(c[0] == 1.f);
   REQUIRE(c[1] == 2.f);
@@ -59,7 +40,26 @@ TEST_CASE("[C] concat_safe()")
   REQUIRE(c[4] == 5.f);
 }
 
-TEST_CASE("[C] raise_error()")
+TEST_CASE("[C] hello__string__to_uppercase()")
+{
+  char *string = hello__string__to_uppercase("Hello World!");
+
+  CHECK(std::string(string) == "HELLO WORLD!");
+
+  hello__string__destroy_string(string);
+}
+
+TEST_CASE("[C] hello__string__to_uppercase_safe()")
+{
+  std::string in_string("Hello World!");
+  std::string out_string(in_string.length(), '\0');
+
+  hello__string__to_uppercase_safe(in_string.c_str(), (char *)out_string.c_str());
+
+  REQUIRE(out_string == "HELLO WORLD!");
+}
+
+TEST_CASE("[C] hello__error__raise_error()")
 {
   int result = hello__error__raise_error();
   REQUIRE(result < 0);
