+++ nodejs/hello/__tests__/test_array.js
@@ -1,0 +1,5 @@
+const hello = require("../hello");
+
+test("concat", () => {
+  expect(hello.array.concat([1, 2], [3, 4, 5])).toEqual([1, 2, 3, 4, 5]);
+});
