--- nodejs/hello/__tests__/test_hello.js
@@ -1,5 +1,0 @@
-const hello = require("../hello");
-
-test("concat", () => {
-  expect(hello.concat([1, 2], [3, 4, 5])).toEqual([1, 2, 3, 4, 5]);
-});
