+++ python/hello/tests/test_array.py
@@ -1,0 +1,4 @@
+import hello
+
+def test_concat():
+    assert hello.array.concat([1, 2], [3, 4, 5]) == [1, 2, 3, 4, 5]
