+++ rust/hello/src/array.rs
@@ -1,0 +1,16 @@
+pub fn concat(a: &[f32], b: &[f32]) -> Vec<f32> {
+    [a, b].concat()
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn test_concat() {
+        assert_eq!(
+            concat(&[1.0, 2.0], &[3.0, 4.0, 5.0]),
+            [1.0, 2.0, 3.0, 4.0, 5.0]
+        );
+    }
+}
